Deduplicate CI scripts and BUILD.bazel
- Extract shared WorkflowConfig/WorkflowManifest to tools/ci/models.py
- Make py_binary cli depend on cli_lib instead of duplicating deps
- Remove temp file usage in bazel-diff, capture stdout directly
