+++ props/critic_dev/BUILD.bazel
@@ -70,21 +70,5 @@ py_binary(
     imports = ["../.."],
     main = "__main__.py",
     visibility = ["//visibility:public"],
-    deps = [
-        ":cli_helpers",
-        "//agent_pkg/runtime",
-        "//cli_util",
-        "//props/cli:cmd_stats",
-        "//props/core:agent_helpers",
-        "//props/core:agent_types",
-        "//props/core:display",
-        "//props/core:eval_client",
-        "//props/core:splits",
-        "//props/core/models:examples",
-        "//props/db:models",
-        "//props/db:session",
-        "@pypi//rich",
-        "@pypi//sqlalchemy",
-        "@pypi//typer",
-    ],
+    deps = [":cli_lib"],
 )
